requirements.txt
#Core numerical and image processing libraries
numpy
scipy
Pillow
matplotlib
imageio

#Configuration and CLI
PyYAML
tqdm

#Testing
pytest